[TM-1435] dashboard tree restoration goal controller moved to job (#541)
* [TM-1435] dashboard tree restoration goal controller moved to job

* [TM-1435] dashboard tree restoration goal controller moved to job
