+++ app/Http/Controllers/V2/Dashboard/ViewTreeRestorationGoalController.php
@@ -2,174 +2,51 @@
 
 namespace App\Http\Controllers\V2\Dashboard;
 
-use App\Helpers\TerrafundDashboardQueryHelper;
 use App\Http\Controllers\Controller;
-use App\Http\Resources\V2\Dashboard\ViewTreeRestorationGoalResource;
-use App\Models\V2\Projects\Project;
-use App\Models\V2\Projects\ProjectReport;
-use App\Models\V2\Sites\Site;
-use App\Models\V2\Sites\SiteReport;
-use App\Models\V2\TreeSpecies\TreeSpecies;
-use Carbon\Carbon;
+use App\Http\Resources\DelayedJobResource;
+use App\Jobs\Dashboard\RunTreeRestorationGoalJob;
+use App\Models\DelayedJob;
+use App\Models\Traits\HasCacheParameter;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Redis;
 
 class ViewTreeRestorationGoalController extends Controller
 {
+    use HasCacheParameter;
+
     public function __invoke(Request $request): JsonResponse
     {
-        $query = $this->prepareProjectQuery($request);
-        $rawProjectIds = $this->getRawProjectIds($query);
-        $allProjectIds = $this->getAllProjectIds($rawProjectIds);
-        $siteIds = $this->getSiteIds($allProjectIds);
-        $distinctDates = $this->getDistinctDates($siteIds);
-        $latestDueDate = $this->getLatestDueDate($distinctDates);
+        try {
+            $cacheParameter = $this->getParametersFromRequest($request);
+            $cacheValue = Redis::get('tree-restoration-goal-'.$cacheParameter);
 
-        $forProfitProjectIds = $this->filterProjectIdsByType($rawProjectIds, 'for-profit-organization');
-        $nonProfitProjectIds = $this->filterProjectIdsByType($rawProjectIds, 'non-profit-organization');
-        $forProfitSiteIds = $this->getSiteIds($forProfitProjectIds);
-        $nonProfitSiteIds = $this->getSiteIds($nonProfitProjectIds);
+            if (! $cacheValue) {
+                $frameworks = data_get($request, 'filter.programmes', []);
+                $landscapes = data_get($request, 'filter.landscapes', []);
+                $organisations = data_get($request, 'filter.organisations.type', []);
+                $country = data_get($request, 'filter.country', '');
 
-        $forProfitTreeCount = $this->treeCountByDueDate($forProfitProjectIds);
-        $nonProfitTreeCount = $this->treeCountByDueDate($nonProfitProjectIds);
+                $delayedJob = DelayedJob::create();
+                $job = new RunTreeRestorationGoalJob(
+                    $delayedJob->id,
+                    $frameworks,
+                    $landscapes,
+                    $organisations,
+                    $country,
+                    $cacheParameter
+                );
+                dispatch($job);
 
-        $totalTreesGrownGoal = $query->sum('trees_grown_goal');
-
-        $treesUnderRestorationActualTotal = $this->treeCountPerPeriod($siteIds, $distinctDates, $totalTreesGrownGoal);
-        $treesUnderRestorationActualForProfit = $this->treeCountPerPeriod($forProfitSiteIds, $distinctDates, $totalTreesGrownGoal);
-        $treesUnderRestorationActualNonProfit = $this->treeCountPerPeriod($nonProfitSiteIds, $distinctDates, $totalTreesGrownGoal);
-
-        $averageSurvivalRateTotal = $this->getAverageSurvival($allProjectIds);
-        $averageSurvivalRateForProfit = $this->getAverageSurvival($forProfitProjectIds);
-        $averageSurvivalRateNonProfit = $this->getAverageSurvival($nonProfitProjectIds);
-
-        $result = [
-            'forProfitTreeCount' => (int) $forProfitTreeCount,
-            'nonProfitTreeCount' => (int) $nonProfitTreeCount,
-            'totalTreesGrownGoal' => (int) $totalTreesGrownGoal,
-            'treesUnderRestorationActualTotal' => $treesUnderRestorationActualTotal,
-            'treesUnderRestorationActualForProfit' => $treesUnderRestorationActualForProfit,
-            'treesUnderRestorationActualNonProfit' => $treesUnderRestorationActualNonProfit,
-            'averageSurvivalRateTotal' => floatval($averageSurvivalRateTotal),
-            'averageSurvivalRateForProfit' => floatval($averageSurvivalRateForProfit),
-            'averageSurvivalRateNonProfit' => floatval($averageSurvivalRateNonProfit),
-        ];
-
-        return new JsonResponse(ViewTreeRestorationGoalResource::make($result));
-    }
-
-    private function prepareProjectQuery(Request $request)
-    {
-        return TerrafundDashboardQueryHelper::buildQueryFromRequest($request);
-    }
-
-    private function getRawProjectIds($query)
-    {
-        return $query
-            ->select('v2_projects.id', 'organisations.type')
-            ->get();
-    }
-
-    private function getAllProjectIds($projectIds)
-    {
-        return $projectIds->pluck('id')->toArray();
-    }
-
-    private function getSiteIds($projectIds)
-    {
-        return Site::whereIn('project_id', $projectIds)->whereIn('status', Site::$approvedStatuses)->pluck('id');
-    }
-
-    private function getDistinctDates($siteIds)
-    {
-        return SiteReport::selectRaw('YEAR(due_at) as year, MONTH(due_at) as month')
-            ->whereNotNull('due_at')
-            ->whereIn('site_id', $siteIds)
-            ->groupBy('year', 'month')
-            ->get()
-            ->toArray();
-    }
-
-    private function filterProjectIdsByType($projectIds, $type)
-    {
-        return collect($projectIds)->filter(function ($row) use ($type) {
-            return $row->type === $type;
-        })->pluck('id')->toArray();
-    }
-
-    private function treeCountByDueDate(array $projectIds)
-    {
-        $projects = Project::whereIn('id', $projectIds)->get();
-
-        return $projects->sum(function ($project) {
-            return $project->trees_planted_count;
-        });
-    }
-
-    private function treeCountPerPeriod($siteIds, $distinctDates, $totalTreesGrownGoal)
-    {
-        $treesUnderRestorationActual = [];
-        $totalAmount = 0;
-
-        foreach ($distinctDates as $date) {
-            $year = $date['year'];
-            $month = $date['month'];
-            $treeSpeciesAmount = 0;
-
-            $reports = SiteReport::whereIn('site_id', $siteIds)
-                ->whereNotIn('v2_site_reports.status', SiteReport::UNSUBMITTED_STATUSES)
-                ->whereYear('v2_site_reports.due_at', $year)
-                ->whereMonth('v2_site_reports.due_at', $month)
-                ->get();
-
-            foreach ($reports as $report) {
-                $treeSpeciesAmount += $report->treeSpecies()->where('collection', TreeSpecies::COLLECTION_PLANTED)->sum('amount');
+                return (new DelayedJobResource($delayedJob))->additional(['message' => 'Data for total-section-header is being processed']);
+            } else {
+                return response()->json(json_decode($cacheValue));
             }
+        } catch (\Exception $e) {
+            Log::error('Error calculating tree restoration goal: ' . $e->getMessage());
 
-            $totalAmount = $totalTreesGrownGoal;
-
-            $formattedDate = Carbon::create($year, $month, 1);
-
-            $treesUnderRestorationActual[] = [
-                'dueDate' => $formattedDate,
-                'treeSpeciesAmount' => (int) $treeSpeciesAmount,
-                'treeSpeciesPercentage' => 0,
-            ];
+            return response()->json(['error' => 'An error occurred while calculating tree restoration goal'], 500);
         }
-
-        foreach ($treesUnderRestorationActual as &$treeData) {
-            $percentage = ($totalAmount != 0) ? ($treeData['treeSpeciesAmount'] / $totalAmount) * 100 : 0;
-            $treeData['treeSpeciesPercentage'] = floatval(number_format($percentage, 3));
-        }
-
-        return $treesUnderRestorationActual;
-    }
-
-    private function getLatestDueDate($distinctDates)
-    {
-        $latestYear = 0;
-        $latestMonth = 0;
-
-        foreach ($distinctDates as $entry) {
-            $year = $entry['year'];
-            $month = $entry['month'];
-
-            if ($year > $latestYear || ($year == $latestYear && $month > $latestMonth)) {
-                $latestYear = $year;
-                $latestMonth = $month;
-            }
-        }
-
-        $latestDate = [
-            'year' => $latestYear,
-            'month' => $latestMonth,
-        ];
-
-        return $latestDate;
-    }
-
-    private function getAverageSurvival(array $projectIds)
-    {
-        return ProjectReport::isApproved()->whereIn('project_id', $projectIds)->avg('pct_survival_to_date');
     }
 }
